+++ src/Stepper/index.ts
@@ -1,3 +1,9 @@
-export { default as Stepper } from './Stepper'
-export { default as useStep } from './useStep'
-export type { OnDone, StepMetadata, State as StepperState, UseStepMethods} from './useStep'
+export { default as Stepper } from './Stepper';
+export { default as useStep } from './useStep';
+export type {
+  OnDone,
+  StepMetadata,
+  State as StepperState,
+  UseStepMethods,
+} from './useStep';
+export type { StepperProps } from './Stepper';