+++ stories/components/Card/argTypes.ts
@@ -1,0 +1,12 @@
+import { VARIANT } from '../../helpers';
+
+export const CARD_COMMON_PROPS = {
+  bg: {
+    control: { type: 'select' },
+    options: VARIANT,
+  },
+  border: {
+    control: { type: 'select' },
+    options: VARIANT,
+  },
+};